refactor: Class names
